+++ src/main/java/org/asamk/signal/commands/DaemonCommand.java
@@ -36,9 +36,6 @@ public class DaemonCommand implements MultiLocalCommand {
         subparser.addArgument("--ignore-attachments")
                 .help("Don’t download attachments of received messages.")
                 .action(Arguments.storeTrue());
-        subparser.addArgument("--json")
-                .help("WARNING: This parameter is now deprecated! Please use the global \"--output=json\" option instead.\n\nOutput received messages in json format, one json object per line.")
-                .action(Arguments.storeTrue());
     }
 
     @Override
@@ -48,12 +45,7 @@ public class DaemonCommand implements MultiLocalCommand {
 
     @Override
     public void handleCommand(final Namespace ns, final Manager m) throws CommandException {
-        var inJson = ns.get("output") == OutputType.JSON || ns.getBoolean("json");
-
-        // TODO delete later when "json" variable is removed
-        if (ns.getBoolean("json")) {
-            logger.warn("\"--json\" option has been deprecated, please use the global \"--output=json\" instead.");
-        }
+        var inJson = ns.get("output") == OutputType.JSON;
 
         boolean ignoreAttachments = ns.getBoolean("ignore-attachments");
 
@@ -82,12 +74,7 @@ public class DaemonCommand implements MultiLocalCommand {
 
     @Override
     public void handleCommand(final Namespace ns, final List<Manager> managers) throws CommandException {
-        var inJson = ns.get("output") == OutputType.JSON || ns.getBoolean("json");
-
-        // TODO delete later when "json" variable is removed
-        if (ns.getBoolean("json")) {
-            logger.warn("\"--json\" option has been deprecated, please use the global \"--output=json\" instead.");
-        }
+        var inJson = ns.get("output") == OutputType.JSON;
 
         boolean ignoreAttachments = ns.getBoolean("ignore-attachments");